finish show more in recommendation
